chore: revert to using `details` instead of `error`
After discussing with @kwasniew, I've decided to keep the `details`
property of the error instead of using the new `errors` property.

Additionally, while I have added the `details[].description` property
intended to replace `details[].message`, I have not removed
`details[].message`. Because this is a breaking API change, this
approach is more in line with Unleash's general
deprecation strategy: when you deprecate something, keep it around for
an extra full major release.

This has the following benefits:
- it avoids the awkward repetition: `error.errors` instead becomes
`error.details`
- if there are any places in the code where we rely on
`error.details[].message`, this will not break because the `message`
field is still present.