chore: enable exhaustiveness in switch lint + no fallthrough cases
This change enables the `tsconfig` option
`noFallthroughCasesInSwitch` and fixes the one instance we had of that
in the code.

It also enables the `switch-exhaustiveness-check` option for the
linter. This seems to at least catch errors where you're switching on
a set of known strings (e.g. `"X" | "Y" | "Z"`), but may not catch all
other kinds, such as switching on arbitrary numbers etc.