docs: ADR: Separation of Request and Response schemas (#3869)
https://linear.app/unleash/issue/2-943/adr-separate-request-and-response-types-in-apis

During the updating of our OpenAPI documentation, we've seen several
times that our schemas are either way too wide for a response or way to
strict for a request. This is usually due to us reusing the same schema
for both request and response. We should write an ADR where we reason
about the usefulness of code duplication and keeping separate response
and request schemas.

Based on our needs, this PR adds my suggested ADR.

---------

Co-authored-by: Simon Hornby <liquidwicked64@gmail.com>