chore: rename version to reflect next candidate (#3944)
## About the changes
Rationale: we're doing continuous deployment of every commit. We don't
want to bump versions with every deployment because it doesn't add
value.

Having a `-beta.xx` is also not relevant now, as any commit is deployed,
but the repository on itself is not published on npm as a different
version

We discussed some options: `rc` for release candidate, `beta`, but we
thought main reflects better that this version does not change unless
there's a reason for changing it (i.e. we're preparing the next version)