Fix: return 400/404 errors if segment id is invalid or doesn't exist
This change adds error handling to the segment store in cases where
the ID isn't a valid postgres integer. This causes it to return a
BadDataError instead of throwing an uncaught Error.

As a bonus, we also catch cases where the row doesn't exist and
populate that error message with more info: it now says that a segment
with <id> doesn't exist instead of just 'No row'.