chore: remove OpenAPI snapshot tests
 Background (general): I don’t think it creates any value. I don’t think people check the snapshot explicitly (especially now that it’s so large that it gets hidden by default), and it only creates an extra chore that you forget to do all the time (at least I do).

Background (why now?): I’ve been fighting a weird issue with the snapshot test for the [API tokens OpenAPI PR](https://github.com/Unleash/unleash/pull/4137). The tests run fine locally, but fail in CI. The tests seem to be using an old version of the snapshot for the test. However, when I check the snapshot in the files tab, it’s got the new and correct values. I’ve spent probably more than an hour trying to fix and debug this today but to no avail.